# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conteudo-padrao:
    get:
      operationId: obterConteudoPadrao
      tags: [conteudo]
      summary: Conteúdo padrão global
      description: Retorna o conteúdo padrão global da página pública.
      responses:
        "200":
          description: Conteúdo padrão
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConteudoPagina"
    put:
      operationId: atualizarConteudoPadrao
      tags: [conteudo]
      summary: Salvar conteúdo padrão global
      description: Substitui o conteúdo padrão global da página pública.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConteudoPagina"
      responses:
        "200":
          description: Conteúdo padrão salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConteudoPagina"
  /pacientes/{id}/conteudo:
    get:
      operationId: obterConteudoPaciente
      tags: [conteudo]
      summary: Conteúdo efetivo da paciente
      description: Retorna o conteúdo da paciente (override se houver, senão o padrão global) e se é personalizado.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conteúdo da paciente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConteudoPaciente"
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
    put:
      operationId: atualizarConteudoPaciente
      tags: [conteudo]
      summary: Salvar override da paciente
      description: Salva um conteúdo personalizado (override) para a paciente.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConteudoPagina"
      responses:
        "200":
          description: Override salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConteudoPaciente"
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
    delete:
      operationId: removerConteudoPaciente
      tags: [conteudo]
      summary: Reverter ao padrão
      description: Remove o override da paciente, voltando ao conteúdo padrão global.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Override removido (volta ao padrão)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConteudoPaciente"
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    Erro:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `application/pdf`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SecaoEtapa:
      type: object
      description: Etapa da linha do tempo. `offsetDias` é relativo à data da cirurgia (0 = no dia, null = sem data). `data` é preenchida só na página pública.
      properties:
        quando:
          type: string
        titulo:
          type: string
        descricao:
          type: string
        offsetDias:
          type: integer
          nullable: true
        data:
          type: string
      required:
        - quando
        - titulo
        - descricao
    SecaoContato:
      type: object
      properties:
        rotulo:
          type: string
        valor:
          type: string
      required:
        - rotulo
        - valor
    SecaoMedicamento:
      type: object
      description: Medicamento a suspender (marca + princípio ativo).
      properties:
        marca:
          type: string
        principio:
          type: string
      required:
        - marca
    SecaoGrupoMedicamentos:
      type: object
      description: Janela de suspensão. `offsetDias` é relativo à data da cirurgia (negativo = antes, null = sem data-limite). `data` (dd/mm) é preenchida só na página pública.
      properties:
        quando:
          type: string
        offsetDias:
          type: integer
          nullable: true
        data:
          type: string
        medicamentos:
          type: array
          items:
            $ref: "#/components/schemas/SecaoMedicamento"
      required:
        - quando
        - medicamentos
    SecaoArquivo:
      type: object
      description: Metadados de arquivo anexado a uma seção. Os bytes vivem no bucket; `token` é opaco e também é a chave do objeto.
      properties:
        nomeArquivo:
          type: string
        tamanho:
          type: integer
        token:
          type: string
      required:
        - nomeArquivo
        - tamanho
        - token
...
    SecaoConteudo:
      type: object
      description: Bloco editável da página pública. O campo `tipo` define quais campos opcionais são usados na renderização.
      properties:
        id:
          type: string
        tipo:
          type: string
          enum:
            - linha_do_tempo
            - lista
            - documentos
            - politica
            - contatos
            - texto
            - preparo
            - suspensao_medicamentos
            - preparo_pele
            - receituario_posop
        titulo:
          type: string
        itens:
          type: array
          items:
            type: string
        corpo:
          type: string
        etapas:
          type: array
          items:
            $ref: "#/components/schemas/SecaoEtapa"
        contatos:
          type: array
          items:
            $ref: "#/components/schemas/SecaoContato"
        grupos:
          type: array
          items:
            $ref: "#/components/schemas/SecaoGrupoMedicamentos"
        aviso:
          type: string
        arquivo:
          $ref: "#/components/schemas/SecaoArquivo"
        produtos:
          type: array
          description: preparo_pele — produtos do preparo da pele (nome, instrução, início e tag).
          items:
            $ref: "#/components/schemas/SecaoProduto"
        medicacoes:
          type: array
          description: receituario_posop — medicações do receituário pós-operatório (nome, instrução, via).
          items:
            $ref: "#/components/schemas/SecaoMedicacao"
      required:
        - id
        - tipo
        - titulo
    SecaoMedicacao:
      type: object
      description: receituario_posop — uma medicação do receituário pós-operatório. Todos os campos aceitam variáveis {{...}}.
      properties:
        nome:
          type: string
          description: 'Nome e dose, ex.: "Cefalexina 500mg".'
        instrucao:
          type: string
          description: 'Posologia, ex.: "Tomar 1 comprimido de 6/6 horas por 7 dias".'
        via:
          type: string
          description: 'Via de uso, exibida em itálico entre parênteses, ex.: "Via oral", "Uso ocular".'
      required:
        - nome
        - instrucao
        - via
    SecaoProduto:
      type: object
      description: preparo_pele — um produto do preparo da pele. Todos os campos aceitam variáveis {{...}}.
      properties:
        nome:
          type: string
          description: 'Nome do produto e marca, ex.: "Blancy TX — Mantecorp".'
        instrucao:
          type: string
          description: 'Instrução de uso, ex.: "Aplicar 1 camada na pele à noite, todos os dias".'
        inicio:
          type: string
          description: 'Quando começar, ex.: "Iniciar 10 dias antes da cirurgia".'
        tag:
          type: string
          description: 'Rótulo curto, ex.: "1 frasco · Uso tópico noturno".'
      required:
        - nome
        - instrucao
        - inicio
        - tag
    ConteudoPagina:
      type: object
      description: Conteúdo editável (padrão global ou override de paciente).
      properties:
        secoes:
          type: array
          items:
            $ref: "#/components/schemas/SecaoConteudo"
      required:
        - secoes
    ConteudoPaciente:
      type: object
      description: Conteúdo efetivo da paciente e se ele é personalizado (override) ou herdado do padrão global.
      properties:
        secoes:
          type: array
          items:
            $ref: "#/components/schemas/SecaoConteudo"
        personalizado:
          type: boolean
      required:
        - secoes
        - personalizado
...